logic to specifically check for a top level h1
