+++ packages/content-conformance/src/rules/__tests__/must-have-h1.test.ts
@@ -1,10 +1,9 @@
 import { testRule } from '../../test/utils'
-
-import mustHaveH1 from '../must-have-h1'
+import topOfFileH1 from '../top-of-file-h1'
 
 describe('must-have-h1', () => {
-  test('Content without H1', () => {
-    testRule(mustHaveH1, [
+  test('Content with level 2 heading', () => {
+    testRule(topOfFileH1, [
       {
         fixture: `## I am not an H1`,
         messages: ['Must have a level 1 heading at the top of the file.'],
@@ -12,8 +11,20 @@ describe('must-have-h1', () => {
     ])
   })
 
-  test('no error', () => {
-    testRule(mustHaveH1, [
+  test('Content with no top level h1', () => {
+    testRule(topOfFileH1, [
+      {
+        fixture: `
+            I am some text
+            # I am not an H1
+        `,
+        messages: ['Must have a level 1 heading at the top of the file.'],
+      },
+    ])
+  })
+
+  test('Content with top level h1', () => {
+    testRule(topOfFileH1, [
       {
         fixture: `# I am an H1`,
         messages: [],
